PR feedback

+++ media/mojo/services/mojo_cdm_service.cc
@@ -24,7 +24,7 @@
 
 #if BUILDFLAG(USE_STARBOARD_MEDIA)
 #include "media/starboard/starboard_cdm.h"
-constexpr unsigned long long kMaxMetricsSize = 1024 * 1024;
+constexpr int kMaxMetricsSize = 1024 * 1024;
 #endif  // BUILDFLAG(USE_STARBOARD_MEDIA)
 
 namespace media {
@@ -243,7 +243,6 @@ void MojoCdmService::OnDecryptorConnectionError() {
 
 #if BUILDFLAG(USE_STARBOARD_MEDIA)
 // TODO(b/432075710) move these into a Starboard CDM extension.
-
 void MojoCdmService::GetMetrics(GetMetricsCallback callback) {
 
   auto* starboard_cdm = static_cast<media::StarboardCdm*>(cdm_.get());
